Fix window not responding after losing focus
After the render context starts the render tasks, it will sit in a loop
handling the window messages. When the user closes the window, the
render context will set a flag that tells the threads to stop. This flag
is checked in each thread before shading a pixel. If any thread exits
before all of the pixels in a block are shaded, another flag will be set
to signal that the image was partially rendered. This prevents an
incomplete image from being saved.

Fix #2